+++ backend/src/server.js
@@ -1,7 +1,6 @@
 const express = require('express');
 const cors = require('cors');
 require('dotenv').config();
-const mariadb = require('mariadb');
 
 const app = express();
 const port = process.env.PORT || 5000;
@@ -9,15 +8,8 @@ const port = process.env.PORT || 5000;
 app.use(cors());
 app.use(express.json());
 
-const pool = mariadb.createPool({
-    host: process.env.DB_HOST,
-    user: process.env.DB_USER,
-    password: process.env.DB_PASSWORD,
-    database: process.env.DB_NAME,
-    connectionLimit: 5
-});
-
 // Test Route for Setup
+/*
 app.get('/api/v1', (req, res) => {
     res.json({message: 'Server is running'});
 });
@@ -36,6 +28,10 @@ app.get('/api/v1/test-db', async (req, res) => {
         if (connection) connection.release();
     }
 });
+*/
+
+const userRoutes = require('./routes/usersRoutes');
+app.use('/api/users', userRoutes);
 
 app.listen(port, () => {
     console.log(`Server is running on http://localhost:${port}`);